seperate test for exp circuit
