+++ circuit-benchmarks/src/exp_circuit.rs
@@ -24,7 +24,7 @@ mod tests {
     use rand::SeedableRng;
     use rand_xorshift::XorShiftRng;
     use zkevm_circuits::evm_circuit::witness::{block_convert, Block};
-    use zkevm_circuits::exp_circuit::test::ExpCircuit;
+    use zkevm_circuits::exp_circuit::TestExpCircuit;
 
     use crate::bench_params::DEGREE;
 
@@ -38,7 +38,7 @@ mod tests {
         let base = Word::from(132);
         let exponent = Word::from(27);
         let block = generate_full_events_block(DEGREE, base, exponent);
-        let circuit = ExpCircuit::<Fr>::new(block);
+        let circuit = TestExpCircuit::<Fr>::new(block);
 
         // Initialize the polynomial commitment parameters
         let mut rng = XorShiftRng::from_seed([
@@ -68,7 +68,7 @@ mod tests {
             Challenge255<G1Affine>,
             XorShiftRng,
             Blake2bWrite<Vec<u8>, G1Affine, Challenge255<G1Affine>>,
-            ExpCircuit<Fr>,
+            TestExpCircuit<Fr>,
         >(
             &general_params,
             &pk,
