+++ zkevm-circuits/src/exp_circuit/param.rs
@@ -1,0 +1,5 @@
+/// The number of rows assigned for each step in an exponentiation trace.
+pub const OFFSET_INCREMENT: usize = 7usize;
+/// The number of rows required for the exponentiation table within the circuit
+/// for each step.
+pub const ROWS_PER_STEP: usize = 4usize;
